+++ cnxarchive/views_folder/extras.py
@@ -1,0 +1,129 @@
+# -*- coding: utf-8 -*-
+# ###
+# Copyright (c) 2013, Rice University
+# This software is subject to the provisions of the GNU Affero General
+# Public License version 3 (AGPLv3).
+# See LICENCE.txt for details.
+# ###
+"""All the views."""
+import os
+import json
+import logging
+from datetime import datetime, timedelta
+from re import compile
+
+import psycopg2
+import psycopg2.extras
+from cnxepub.models import flatten_tree_to_ident_hashes
+from lxml import etree
+from pytz import timezone
+from pyramid import httpexceptions
+from pyramid.settings import asbool
+from pyramid.threadlocal import get_current_registry, get_current_request
+from pyramid.view import view_config
+
+from .. import config
+from .. import cache
+# FIXME double import
+from .. import database
+from ..database import SQL, get_tree, get_collated_content
+from ..search import (
+    DEFAULT_PER_PAGE, QUERY_TYPES, DEFAULT_QUERY_TYPE,
+    Query,
+    )
+from ..sitemap import Sitemap
+from ..robots import Robots
+from ..utils import (
+    COLLECTION_MIMETYPE, IdentHashSyntaxError,
+    IdentHashShortId, IdentHashMissingVersion,
+    portaltype_to_mimetype, slugify, fromtimestamp,
+    join_ident_hash, split_ident_hash, split_legacy_hash
+    )
+from .content import _get_content_json
+from .content import get_content_metadata
+from ..views import LEGACY_EXTENSION_MAP
+from .exports import get_export_allowable_types
+
+PAGES_TO_BLOCK = [
+    'legacy.cnx.org', '/lenses', '/browse_content', '/content/', '/content$',
+    '/*/pdf$', '/*/epub$', '/*/complete$',
+    '/*/offline$', '/*?format=*$', '/*/multimedia$', '/*/lens_add?*$',
+    '/lens_add', '/*/lens_view/*$', '/content/*view_mode=statistics$']
+
+logger = logging.getLogger('cnxarchive')
+
+class ExportError(Exception):
+    """Used as catchall for other export errors."""
+
+    pass
+
+
+# #################### #
+#   Helper functions   #
+# #################### #
+
+
+def get_latest_version(uuid_):
+    settings = get_current_registry().settings
+    with psycopg2.connect(settings[config.CONNECTION_STRING]) as db_connection:
+        with db_connection.cursor() as cursor:
+            cursor.execute(SQL['get-module-versions'], {'id': uuid_})
+            try:
+                return cursor.fetchone()[0]
+            except (TypeError, IndexError,):  # None returned
+                raise httpexceptions.HTTPNotFound()
+
+
+def is_latest(id, version):
+    """Determine if this is the latest version of this content."""
+    return get_latest_version(id) == version
+
+
+# ######### #
+#   Views   #
+# ######### #
+
+
+@view_config(route_name='content-extras', request_method='GET')
+def get_extra(request):
+    """Return information about a module / collection that cannot be cached."""
+    settings = get_current_registry().settings
+    exports_dirs = settings['exports-directories'].split()
+    args = request.matchdict
+    id, version = split_ident_hash(args['ident_hash'])
+    results = {}
+
+    with psycopg2.connect(settings[config.CONNECTION_STRING]) as db_connection:
+        with db_connection.cursor() as cursor:
+            results['downloads'] = \
+                list(get_export_allowable_types(cursor, exports_dirs,
+                                                id, version))
+            results['isLatest'] = is_latest(id, version)
+            results['canPublish'] = database.get_module_can_publish(cursor, id)
+
+    resp = request.response
+    resp.content_type = 'application/json'
+    resp.body = json.dumps(results)
+    return resp
+
+
+@view_config(route_name='extras', request_method='GET')
+def extras(request):
+    """Return a dict with archive metadata for webview."""
+    settings = get_current_registry().settings
+    with psycopg2.connect(settings[config.CONNECTION_STRING]) as db_connection:
+        with db_connection.cursor() as cursor:
+            metadata = {
+                'languages_and_count': _get_available_languages_and_count(
+                    cursor),
+                'subjects': list(_get_subject_list(cursor)),
+                'featuredLinks': _get_featured_links(cursor),
+                'messages': _get_service_state_messages(cursor),
+                'licenses': _get_licenses(cursor),
+                }
+
+    resp = request.response
+    resp.status = '200 OK'
+    resp.content_type = 'application/json'
+    resp.body = json.dumps(metadata)
+    return resp
